Move v2 and v1 uri features into the correct modules
The uri tests were looking at bip21 uris with v1 or v2 specific features
and failing when the features were locked to the opposite feature
respectively.

I have moved the relevant tests to be in the correct module so that they
only compile within the correct feature.
